ADD: arrow btn tp service plugin (#66)
* change to vue3

without any errors

* new Update

it's a last dbug, with the better css

* Update OneClick.vue

page of component one click is created, with the style and template. scripts are not update

* new update

some new update with dbug

* new update for node page

node page added without content but in the route installed

* make ui

make a Trapezium on ui design and add this to the main js for use every wher

* update for the Node page

it's just simple page without any component. it has a small bug

* nodeBg added

the node page's back ground added to main.js. it has some css problems at he moment but it's ok. the routing at the moment is symple. without any conditions! 😏

* update oneclick with v-for

added v-for and data to the row but they are not inline block

* update

* update

* node header

a component added to header for the node page

* node header js

node header as a base component

* header added

the function is good but css has problem, positioning is not true

* solve the css problems

header with the logo fixed on node page

* update

update the css for header in the node page

* installation option stracture

one click changed

* update installOption

css solved

* right side bar of node page component added

* update

* update

* update

* right side bar added to node with dummy style

* update

* update

* update

* update

* update

* update

* update

* FE Update

* new plugin for node page

new plugin made and connected tgo node page
footerNode
ServicePlugin
theTrapezion

* update

* gridnode added

add a grid page to be a cabinet of node page

* update

stracture of node page is ready

* control update

control panel
control grid
routing added

* update

* conflicts solved

* update fixed lockjson

* jest update

* trapezium positions

* update

* update

* update

* update

* update

* nodedrag added

* colors set to design

* ADD:Service-plugin_style

* update

* ADD:arrow btn to service plugin