ADJUST: UI on ControlPage, MEV-Boost Relays and Documentation (#1053)
* added support modal box comp. with dynamic image

* Support modal now displays dynamic documentation link based on current network

In this commit, I have updated the support modal to dynamically display the appropriate documentation link based on the current network. The documentation links for mainnet, testnet, and gnosis have been added to the code, and the modal now opens the correct link based on the value of the current network. This improves the user experience by providing quick and easy access to relevant information for their specific network.

* Support modal text now dynamic with i18n for internationalization

In this commit, I have implemented internationalization (i18n) for the support modal text. The modal text is now dynamically generated based on the user's language preference, providing a better experience for users speaking different languages.

* Removed unnecessary code to make component more clean and efficient

* feat: update panel auto update status

- dynamic text size fixed
- auto update has the on off status with color
 Fixes #1040

* update: update panel in the i18n updated

* feat: shuffle available block relays in mevBoostConfig component

* feat: shuffle available block relays in ModifyPanel component

* @mabasian feat: shuffle available block relays in addPanel component

* update: the if else statement to switch case for more readable for dev

* Added circular progress bar component for syncing status in control page

* feat: shuffle available block relays in MevboostModal component

* Refactor syncstatus component

* Dynamic color for circular progress bar

* update color piker for progress circular

* syncstatus return to the old form ready for wiring