Refactor Log Access System and Fix Critical CSS Bugs (#1066)
* Fix container style for tester contributors

* Remove extraneous files

Some files were found to be no longer needed in the project and were deleted from the control file.
Removing these files also helps to ensure that only necessary files are included in the Control page and reduces the risk of introducing any unintended changes to the codebase.

This commit serves to document the removal of these files from the control file.

* update the diktat for slashing in the i18n

* Log Navigation UI is complete

The Log Navigation UI has been fully implemented and is now functional. The design and layout have been finalized, and all the necessary components have been added.

* Completed log navigation functionality with placeholder icons.

* Add Log Navigation to NodeJournal

Implemented log navigation feature in NodeJournal. Logs can now be easily accessed and navigated, making it easier to identify and troubleshoot issues.

* Keep Selected Relays on Top in MevBoost

Implemented the feature to keep selected relays at MevBoost always on top. The system now ensures that the selected relays are always displayed at the top of the available blocs, providing a more efficient and user-friendly experience. The changes were made to the available blocs, making it easier for users to quickly access and manage their selected relays.